Fix build script
I was able to use `pump` instead of pipes to get more details on the real error during all of the streaming. Uglify doesn't support minifying ES6. I am not familiar with Browserify, but I think it is pulling in node libraries that have `const` in them, and it isn't transforming those with babel, just the source code. Maybe an older version of browserify or maybe node(?) would allow this build to work without this patch.

Since we are building a chrome extension, const and other ES6 is fine to leave in the production code. uglify-es is now used for minification. It handles ES6.